README.md
This is a sheet that will introduce the a very basic introduction to numpy for data analysis. I have worked on it as a part of taking a course Python Data Analysis from University of Michigan through Coursera. I am extremely obliged to Coursera and University of Michigan in providing such an elaborate content.